Arrumando app.js

+++ src/views/users.php
@@ -32,13 +32,12 @@
                             class="btn btn-warning rounded-circle mr-2">
                             <i class="icofont-edit"></i>
                         </a>
-                        <a href="?delete=<?= $user->id ?>"
-                            class="btn btn-danger rounded-circle">
+                        <a href="?delete=<?= $user->id ?>" class="btn btn-danger rounded-circle" data-confirm>
                             <i class="icofont-trash"></i>
                         </a>
                     </td>
                 </tr>
-            <?php endforeach?>
+            <?php endforeach ?>
         </tbody>
     </table>
 </main>
